+++ src/plugin.hpp
@@ -13,3 +13,4 @@ extern Model* modelCosmicClock;
 extern Model* modelCoronalAnnihilator;
 extern Model* modelLitanyEngine;
 extern Model* modelSyzygy;
+extern Model* modelEternalEclipse;